+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+
+* Fully refactored the library.
+* Removed `SliverTableController`, so you can now pass all arguments directly to the `SliverTable`
+  constructor.
+* Added `rowContainerBuilder` and `topHeaderContainerBuilder`.
+
 ## 0.0.1
 
 * Initial release.